requirements.txt

yfinance>=0.2.40
pandas>=2.2.2
numpy>=1.26
pyyaml>=6.0.2
rich>=13.7
streamlit>=1.30.0
matplotlib>=3.8.0